Fix casting warning

+++ src/loader.cpp
@@ -26,7 +26,7 @@ Pack2::Pack2(std::filesystem::path path_, std::span<uint8_t> data): buf_(data), 
     logger::info("Loading {} assets from {}{}", asset_count(), name, path.extension().string());
     const std::span<uint8_t> asset_data = buf_.subspan(map_offset(), asset_count() * sizeof(Asset2Raw));
     assets = std::span<Asset2Raw>((Asset2Raw*)asset_data.data(), asset_count());
-    for(uint64_t i = 0; i < assets.size(); i++) {
+    for(uint32_t i = 0; i < assets.size(); i++) {
         logger::debug("assets[{}].name_hash = 0x{:016x}", i, assets[i].name_hash);
         namehash_to_asset[assets[i].name_hash] = i;
     }
